requirements.txt
Django==2.2.6
pytz==2019.3
sqlparse==0.3.0

#Instalaciones para rest-framework y heroku
dj-database-url==0.5.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
Markdown==3.0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-decouple==3.1
whitenoise==4.1.2
